This commit represents a set of changes and patches required to overcome aggressive code warnings and fail on warnings that are being triggered as a result of Yocto Project's upgrade to the GCC 11.1 compiler.
Most patches condition on whether or not the -DCMAKE_BUILD_TYPE=Release flag is thrown.

The exception to this is a patch in aws-iot-device-client where the compiler found that the C++ thread header was not included so the patches include the header file.  See the patchfile for that recipe for more details.