recipes-sdk: do not overwrite/clear PACKAGES variable
Clearing PACKAGES variable to only contain ${PN} is very bad and it's not clear
what the purpose is. It doesn't cause much trouble during the regular build for
the target, as do_populate_sysroot() is able to collect required artifacts. But
since there are no ${PN}-dev, ${PN}-dbg and ${PN}-staticdev packages produced,
it renders an SDK with AWS IoT components completely useless, especially since
majority of them only produce .a libraries.

This change removes overwriting/clearing of the PACKAGES variable to use the
default value. If it was done as a workaround for some other issue, a better
solution should be found and not this "sledgehammer approach".

Signed-off-by: Denys Dmytriyenko <denis@denix.org>