qa: cleanup buildspec for layercheck
Remove an option that is pulling in layercheck for meta-python. This is
out of scope for this check and instead we should rely on upstream QA.

Break the command into multiple lines to make it easier to read.

Remove a second `-d` option.

Add --no-auto, --no-auto-dependency, and convert to longer option names.