qa: update buildspec for gatesgarth
Convert `$HOME/dist/meta-aws` to `$CODEBUILD_SRC_DIR`

Remove an option that is pulling in layercheck for meta-python. This is
out of scope for this check and instead we should rely on upstream QA.

Break the command into multiple lines to make it easier to read.

Remove a second `-d` option.

Add --no-auto, and convert to longer option names.

--no-auto-dependency does not exist on gatesgarth