recipes: fix CMAKE_PREFIX_PATH
For recipes that depend on aws-c-common, do_configure fails
when multilib is enabled.

For example, add the following lines to local.conf and then
run 'bitbake aws-c-sdkutils' and we'll see the error:

  MACHINE ?= "qemux86-64"
  require conf/multilib.conf
  MULTILIBS ?= "multilib:lib32"
  DEFAULTTUNE:virtclass-multilib-lib32 ?= "core2-32"

This is because the CMAKE_PREFIX_PATH is not set correctly.
See the search precedure in the link below:
https://cmake.org/cmake/help/latest/command/find_package.html#search-procedure

According to the above link, aws-c-common-config.cmake cannot be
found when CMAKE_PREFIX_PATH is set to ${STAGING_LIBDIR}. The reason
that things still work when multilib is disabled might be that "lib"
is specially handled, as said in the above link, "Always searched."

We'll have other dependencies to search, e.g., s2n. So instead of
changing CMAKE_PREFIX_PATH to ${STAGING_LIBDIR}/cmake, we'll need
to also keep ${STAGING_LIBDIR} in it.

Use aws-c-io as an example, its s2n and aws-c-common's cmake config
files are in the following locations:

  recipe-sysroot/usr/lib64/s2n/cmake/s2n-config.cmake
  recipe-sysroot/usr/lib64/cmake/aws-c-common/aws-c-common-config.cmake

With this patch, find_package will be able to handle cmake config
files in more locations and will be able to build these recipes w/o
multilib enabled.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
(cherry picked from commit 18bb4454d22dc0b475c8d06179aefc5d94acc1ee)