revert UNPACKDIR to WORKDIR as this is breaking nightly updates
devtool is used in a nightly GH workflow to perform checks if new versions
of recipe software is available, this is broken if UNPACKDIR is used.
See more here: https://lists.openembedded.org/g/openembedded-core/topic/112431116#msg215739

Plan is not to backport this to release branches for now. Just for master to
enable auto updates again.