aws-crt-python: update ptest shebang to use python3
Some test scripts in ptest still use `#!/usr/bin/python`, which may not
be available on modern systems where only `python3` is installed.
This change updates the shebang line in all ptest test scripts to use
`#!/usr/bin/python3` to ensure compatibility.

Affected files include (but are not limited to):
  - aws-lc/util/generate-asm-lcov.py
  - s2n/tests/sidetrail/bin/bpl_trace_to_c.py

Signed-off-by: Haixiao Yan <haixiao.yan.cn@windriver.com>