greengrass-*component.bbclass: remove yaml for anonymous functions
It seems that anonymous python functions are executed before any dependent
packages (python3-pyyaml-native) are build. This was discovered by yocto-check-layer
in the yocto autobuilder. Not able to reproduce with Ubuntu 24.04.

The workaround is now to have a simple key value parser not require yaml to read
config files, to set variables.